+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- improve logging: add protocol or node_id to connection logs
+- improve logging: log ignored AT+ or MODBUS commands
+- fix typo in docker-compose.yaml
 - trace heartbeat and regular modbus pakets witl log level DEBUG
 - GEN3PLUS: don't forward ack paket from tsun to the inverter
 - add allow and block filter for AT+ commands